Password system partially implemented

+++ app.py
@@ -1,6 +1,6 @@
 
 import os
-from flask import Flask, render_template, request, redirect
+from flask import Flask, render_template, request, redirect, session, url_for
 from flask_socketio import SocketIO, send, emit
 from flask_sqlalchemy import SQLAlchemy
 
@@ -12,17 +12,24 @@ import models
 
 #db.drop_all()
 #db.create_all()
-
+#i=models.Info("KarnerBlue")
+#db.session.add(i)
+#db.session.commit()
+			
 socketio = SocketIO(app)
 
-		
-@app.route('/')
+@app.route('/', methods=['POST','GET'])		
 def home_page():
-	buses=models.getAllBuses()
+	messages=[]
+	if 'password' in request.form:
+		i=models.Info.query.filter_by(id=1).first()
+		#if submitted password matches password in database
+		if request.form['password'] == i.password:
+			return render_template('main.html',buses=models.getAllBuses())
+		else:
+			messages.append("Incorrect password. Please try again.")
+	return render_template('index.html', messages=messages)
 	
-	return render_template("index.html", buses=buses)
-
-
 @socketio.on('new_message')
 def handle_new_message(message):
     print("New message recieved: ", message)
@@ -32,19 +39,36 @@ def handle_new_message(message):
 
 @app.route('/admin', methods=['POST','GET'])
 def admin():
+	messages=[]
 	if request.method=='POST':
-		if 'newBus' in request.form:
-			b=models.Bus(request.form['newBus'], 'unarrived')
-			db.session.add(b)
-			db.session.commit()
+		
+		if 'newPassword' in request.form:
+			if len(request.form['newPassword']) != 0:
+				i=models.Info.query.filter_by(id=1).first()
+				i.password=request.form['newPassword']
+				db.session.add(i)
+				db.session.commit()
+				messages.append("Password changed to "+request.form['newPassword'])
+				#return redirect(url_for('admin'))
 			
+		
+		if 'newBus' in request.form:
+			if len(request.form['newBus']) != 0:
+				b=models.Bus(request.form['newBus'], 'unarrived')
+				db.session.add(b)
+				db.session.commit()
+				messages.append(request.form['newBus']+" added.")
+				#return redirect(url_for('admin'))
 		if 'busToDelete' in request.form:
-			print("Bus to delete: "+request.form['busToDelete'])
-			b=models.Bus.query.filter_by(busName=request.form['busToDelete']).first()
-			db.session.delete(b)
-			db.session.commit()
+			if len(request.form['busToDelete']) != 0:
+				print("Bus to delete: "+request.form['busToDelete'])
+				b=models.Bus.query.filter_by(busName=request.form['busToDelete']).first()
+				db.session.delete(b)
+				db.session.commit()
+				messages.append(request.form['busToDelete']+" has been deleted.")
+				#return redirect(url_for('admin'))
 	buses=models.getAllBuses()		
-	return render_template('admin.html', buses=buses)
+	return render_template('admin.html', buses=buses, messages=messages)
 	
 @app.route('/updateDatabase', methods=['POST','GET'])
 def updateDatabase():
@@ -58,7 +82,7 @@ def updateDatabase():
 	db.session.add(b)
 	db.session.commit()
 	buses=models.getAllBuses()
-	return render_template('index.html', buses=buses)
+	return render_template('main.html', buses=buses)
 
 if __name__ == '__main__':
     socketio.run(app, host="https://agile-sea-57808.herokuapp.com")
